embed: clean up structured logging calls
