+++ embed/config_logging_journald_windows.go
